+++ App.jsx
@@ -1,0 +1,55 @@
+/**
+ * Sample React Native App
+ * https://github.com/facebook/react-native
+ *
+ * @format
+ * @flow strict-local
+ */
+
+import React from "react";
+import {
+  SafeAreaView,
+  StatusBar,
+  StyleSheet,
+  Text,
+  useColorScheme,
+  View,
+} from "react-native";
+
+import { Colors } from "react-native/Libraries/NewAppScreen";
+
+const styles = StyleSheet.create({
+  container: {
+    alignItems: "center",
+    height: "100%",
+    justifyContent: "center",
+  },
+  title: {
+    fontSize: 24,
+    fontWeight: "600",
+    color: "black",
+    textAlign: "center",
+  },
+});
+
+function App() {
+  const isDarkMode = useColorScheme() === "dark";
+
+  const backgroundStyle = {
+    backgroundColor: isDarkMode ? Colors.darker : Colors.lighter,
+  };
+
+  return (
+    <SafeAreaView style={backgroundStyle}>
+      <StatusBar
+        barStyle={isDarkMode ? "light-content" : "dark-content"}
+        backgroundColor={backgroundStyle.backgroundColor}
+      />
+      <View style={styles.container}>
+        <Text style={styles.title}>Welcome to Children Playground App</Text>
+      </View>
+    </SafeAreaView>
+  );
+}
+
+export default App;
