--- __tests__/App-test.js
@@ -1,14 +1,0 @@
-/**
- * @format
- */
-
-import "react-native";
-import React from "react";
-import renderer from "react-test-renderer";
-import App from "../App";
-
-// Note: test renderer must be required after react-native.
-
-it("renders correctly", () => {
-  renderer.create(<App />);
-});
